requirements.txt
apache-airflow-providers-amazon==3.0.0
sqlparse==0.4.2